Reorganized old route_dynamics module into route_elevation submodule
